Block access to the login.php page when the Session Manager is enabled
Whenever the login modal is enabled, the customer will be redirected back to the home page when he tries to access the login.php page and also after the logout.
